Strict warnings in the unit tests. This commit needs to be carefully reviewed.

+++ libuavcan/include/uavcan/util/lazy_constructor.hpp
@@ -66,7 +66,7 @@ public:
     LazyConstructor(const LazyConstructor<T>& rhs)   // Implicit
         : ptr_(NULL)
     {
-        fill(data_.pool, data_.pool + sizeof(T), 0);
+        fill(data_.pool, data_.pool + sizeof(T), uint8_t(0));
         if (rhs)
         {
             construct<const T&>(*rhs);  // Invoke copy constructor
